Reapplied all changed from the very start, so maybe we get rid of our stability issues

+++ src/com/bulletphysics/collision/narrowphase/DiscreteCollisionDetectorInterface.java
@@ -17,7 +17,7 @@ import javax.vecmath.Vector3d;
 public abstract class DiscreteCollisionDetectorInterface {
 
 	public static abstract class Result {
-		// setShapeIdentifiers provides experimental support for per-triangle material / custom material combiner
+		///setShapeIdentifiers provides experimental support for per-triangle material / custom material combiner
 		public abstract void setShapeIdentifiers(int partId0, int index0, int partId1, int index1);
 
 		public abstract void addContactPoint(Vector3d normalOnBInWorld, Vector3d pointInWorld, double depth);
@@ -27,14 +27,13 @@ public abstract class DiscreteCollisionDetectorInterface {
 		public final Transform transformA = new Transform();
 		public final Transform transformB = new Transform();
 		public double maximumDistanceSquared;
-		//btStackAlloc* m_stackAlloc;
 
 		public ClosestPointInput() {
 			init();
 		}
 		
 		public void init() {
-			maximumDistanceSquared = Double.MAX_VALUE;
+			maximumDistanceSquared = Float.MAX_VALUE;
 		}
 	}
